## Description

<!-- Describe the goal and purpose of this PR. -->

Updates Optimize to use ES Java client 8.16.6. Will not be backported,
and is part of epic to increase our minimum supported DB version

## Checklist

<!--- Please delete options that are not relevant. Boxes should be
checked by reviewer. -->
- [ ] Enable backports when necessary (fex. [for bug
fixes](https://github.com/camunda/camunda/blob/main/CONTRIBUTING.md#backporting-changes)
or [for CI
changes](https://github.com/camunda/camunda/wiki/CI-&-Automation#when-to-backport-ci-changes)).

## Related issues

closes #35606

+++ optimize/util/optimize-commons/src/main/java/io/camunda/optimize/service/db/es/schema/ElasticSearchIndexSettingsBuilder.java
@@ -50,7 +50,8 @@ public class ElasticSearchIndexSettingsBuilder {
                                 n.limit(
                                     configurationService
                                         .getElasticSearchConfiguration()
-                                        .getNestedDocumentsLimit()))));
+                                        .getNestedDocumentsLimit()
+                                        .longValue()))));
   }
 
   public static IndexSettings buildAllSettings(
@@ -78,7 +79,8 @@ public class ElasticSearchIndexSettingsBuilder {
                               n.limit(
                                   configurationService
                                       .getElasticSearchConfiguration()
-                                      .getNestedDocumentsLimit())));
+                                      .getNestedDocumentsLimit()
+                                      .longValue())));
           try {
             indexMappingCreator.getStaticSettings(i, configurationService);
           } catch (final IOException e) {
